reorganize fetch mock code
