+++ test/HostedLayer.test.ts
@@ -1,4 +1,6 @@
-import { test, expect } from 'vitest';
+import { describe, expect, vi, beforeAll, beforeEach } from 'vitest';
+import { customTest as test } from './BaseTest.test';
 import { HostedLayer } from '../src/HostedLayer';
 
 test('Cannot be instantiated directly.', () => {});
+test('Accepts a `token` parameter for authentication.', () => {});
